Update Actuator.cpp
still need to know what to do with values gotten for each actuator so that we can actually move them that amount. however, in this the framework for how the class is set up has been changed to using methods rather than cases. Additionally, the setters and getters were better set up and the main() function that caused an error in the first place was removed entirely.